requirements.txt
uvicorn
fastapi
pygame
bcrypt
pymongo
requests
neat-python